updated console to follow sdc patterns
