--- console/templates/layout.html
@@ -1,53 +1,0 @@
-{% set navigation_bar = [
-    ('/decrypt', 'decrypt', 'Decrypt'),
-    ('/store', 'store', 'Store'),
-] -%}
-
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <title>SDX Console</title>
-    <link rel="stylesheet" href="/static/css/bootstrap.min.css">
-    <script src="/static/js/bootstrap.min.js"></script>
-  </head>
-
-  <body>
-    <div class='container'>
-      <nav class="navbar navbar-default">
-        <div class="container-fluid">
-          <div class="navbar-header">
-            <button type="button" class="navbar-toggle collapsed" data-toggle="collapse" data-target="#navbar" aria-expanded="false" aria-controls="navbar">
-              <span class="sr-only">Toggle navigation</span>
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-              <span class="icon-bar"></span>
-            </button>
-            <a class="navbar-brand" href="/">sdx-console</a>
-          </div>
-          <div id="navbar" class="navbar-collapse collapse">
-            <ul class="nav navbar-nav">
-              {% for href, id, caption in navigation_bar %}
-              <li{% if id == active_page %} class="active" {% endif %}>
-                <a href="{{ href|e }}">{{ caption|e }}</a>
-              </li>
-              {% endfor %}
-              {% if 'Admin' in current_user.roles %}
-                 <li{% if active_page=="AddUser" %} class="active" {% endif %}>
-                     <a href="{{ url_for('add_user') }}">Add New User</a>
-                 </li>
-              {% endif %}
-              {% if current_user.is_authenticated%}
-                <li><a href="{{ url_for('security.logout') }}">Logout</a> </li>
-              {% endif %}
-            </ul>
-          </div>
-        </div>
-      </nav>
-
-      <div class="row">
-        {% block body %}{% endblock %}
-      </div>
-
-    </div>
-  </body>
-</html>
