+++ console/templates/layouts/_base.html
@@ -1,0 +1,77 @@
+<!doctype html>
+<!--[if lt IE 7]><html lang="en-gb" dir="ltr" class="no-js lt-ie9 lt-ie8 lt-ie7"><![endif]-->
+<!--[if IE 7]><html lang="en-gb" dir="ltr" class="no-js lt-ie9 lt-ie8"><![endif]-->
+<!--[if IE 8]><html lang="en-gb" dir="ltr" class="no-js lt-ie9"><![endif]-->
+<!--[if gt IE 8]><!--> <html lang="en-gb" dir="ltr" class="no-js"> <!--<![endif]-->
+  <head>
+    {%- block head %}
+    <meta charset="utf-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge"/>
+    <title>{% block page_title %}{{ _('Survey Data Collection') }}{% endblock page_title %}</title>
+    <meta content="" name="description">
+    <meta content="width=device-width, initial-scale=1" name="viewport" />
+    <link rel="apple-touch-icon" sizes="180x180" href="{{ url_for('static', filename='images/favicons/apple-touch-icon.png') }}" />
+    <link rel="icon" type="image/png" href="{{ url_for('static', filename='images/favicons/favicon-32x32.png') }}" />
+    <link rel="icon" type="image/png" href="{{ url_for('static', filename='images/favicons/favicon-16x16.png') }}" />
+    <link rel="manifest" href="{{ url_for('static', filename='images/favicons/manifest.json') }}" />
+    <link rel="mask-icon" href="{{ url_for('static', filename='images/favicons/safari-pinned-tab.svg') }}" color="#5bbad5" />
+
+    <link rel="shortcut icon" href="{{ url_for('static', filename='images/favicons/favicon.ico') }}" />
+
+    <!--[if lt IE 9]><script src="{{ url_for('static', filename='js/vendor/html5shiv.min.js') }}"></script><![endif]-->
+    <!--[if gt IE 8]><!-->
+      <script>document.documentElement.className = document.documentElement.className.replace('no-js','has-js')</script>
+    <!--<![endif]-->
+    {%- block styles %}
+    {% set css_responsive = "css/"~_theme~"/responsive.css" %}
+    {% set css_fixed = "css/"~_theme~"/fixed.css" %}
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css">
+
+    <!--[if (gt IE 9) | (IEMobile)]><!-->
+    {%- if request.args.get('fixed') == 'true' %}
+      <link href="{{ url_for('static', filename=css_fixed) }}" rel="stylesheet" />
+    {%- else %}
+      <link href="{{ url_for('static', filename=css_responsive) }}" rel="stylesheet" />
+    {%- endif %}
+    <!--<![endif]-->
+    <!--[if (lte IE 9) & (!IEMobile)]>
+      <link href="{{ url_for('static', filename=css_fixed) }}" rel="stylesheet">
+    <![endif]-->
+    <![endif]-->
+    {% endblock styles -%}
+    {% endblock head -%}
+    <link href="{{ url_for('static', filename='site.min.css') }}" rel="stylesheet" />
+
+    <script src="https://code.jquery.com/jquery-3.1.1.js"
+            integrity="sha256-16cdPddA6VdVInumRGo6IbivbERE8p7CQR3HzTBuELA="
+            crossorigin="anonymous"></script>
+
+  </head>
+
+  <body>
+    <div class="page">
+      <div class="page__content">
+
+        <div class="page__header">
+          {% include 'partials/header.html' %}
+        </div>
+
+        <div class="container page__container">
+          {%- block content %}{% endblock content -%}
+        </div>
+      </div>
+      <div class="page__footer">
+        {% include 'partials/footer.html' %}
+      </div>
+
+    </div>
+
+    <!--[if gte IE 9]><!-->
+    {% block scripts %}
+      <script src="{{ url_for('static', filename='js/bundle.js') }}"></script>
+    {% endblock scripts %}
+  	<!--<![endif]-->
+
+  </body>
+
+</html>
