+++ console/templates/layouts/_onecol.html
@@ -1,0 +1,13 @@
+{% extends 'layouts/_base.html' %}
+
+{% block content %}
+
+<div class="grid">
+  <div class="grid__col col-12@xs">
+    <div role="main" class="page__main" id="main">
+      {% block main %}{% endblock %}
+    </div>
+  </div>
+</div>
+
+{% endblock %}
