+++ console/templates/logout.html
@@ -1,7 +1,7 @@
 {% import 'partials/section.html' as section %}
 {% extends "layouts/_twocol.html" %}
 
-{% block page_title %} Home {% endblock %}
+{% block page_title %} Logout {% endblock %}
 
 
 {% block content %}
@@ -14,8 +14,9 @@
       {% if current_user.is_authenticated %}
         <h3>Logged in as: {{current_user.email}}</h3>
       {% else %}
+        You have been logged out.<br>
         To use this service you need to
-        <a href="{{ url_for('security.login', next=request.path)}}"> log in</a>
+        <a href="{{ url_for('security.login', next='home')}}"> log in</a>
       {% endif %}
     </div>
   </div>
