+++ console/templates/partials/footer.html
@@ -1,0 +1,57 @@
+<footer class="footer page__footer" role="contentinfo">
+    <div class="container">
+        <div class="grid">
+            <div class="grid__col col-4@m">
+                <p class="venus footer__heading"><strong>Legal information</strong></p>
+                <ul class="list footer__list">
+                    <li>
+                        <a href="/cookies-privacy"
+                           class="footer__link">Cookies and privacy</a>
+                    </li>
+                </ul>
+            </div>
+
+            <div class="grid__col col-4@m">
+                <p class="venus footer__heading"><strong>About ONS</strong></p>
+                <ul class="list footer__list">
+                    <li>
+                        <a href="https://www.ons.gov.uk/aboutus/whatwedo"
+                           class="footer__link">What we do</a>
+                    </li>
+                    <li>
+                        <a href="/contact-us"
+                           class="footer__link">Contact us</a>
+                    </li>
+                    <li>
+                        <a href="https://www.ons.gov.uk/help/accessibility"
+                           class="footer__link">Accessibility</a>
+                    </li>
+                </ul>
+            </div>
+
+            <div class="grid__col col-4@m">
+                <p class="venus footer__heading"><strong>Statistics</strong></p>
+                <ul class="list footer__list">
+                    <li>
+                        <a href="https://www.statisticsauthority.gov.uk/about-the-authority/"
+                           class="footer__link">UK Statistics Authority</a>
+                    </li>
+                    <li>
+                        <a href="https://www.ons.gov.uk/releasecalendar"
+                           class="footer__link">Release calendar</a>
+                    </li>
+                    <li>
+                        <a href="https://www.ons.gov.uk/news"
+                           class="footer__link">News</a>
+                    </li>
+                </ul>
+            </div>
+
+            <div class="grid__col col-12@m footer__license">
+                <img alt="OGL" class="ogl-icon" src="/static/images/ogl.svg"> All content is available under the <a
+                    href="https://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/">Open Government
+                    Licence v3.0</a>, except where otherwise stated
+            </div>
+        </div>
+    </div>
+</footer>
