+++ console/templates/partials/header.html
@@ -1,0 +1,29 @@
+
+<div id="helpinfo" class="helpinfo">
+  <div class="header__help js-help-body first-closed">
+    <div class="header__container container">
+
+    </div>
+  </div>
+</div>
+
+<div class="container">
+  <header class="header">
+
+    <div class="logo header__logo">
+      <img src="{{ url_for('static', filename='images/logo.svg') }}" alt="Office for National Statistics"
+             class="logo__img"/>
+    </div>
+
+    <button aria-controls="helpinfo" class="header__btn btn btn--light btn--border u-vh js-help-btn">
+      Help
+    </button>
+
+  </header>
+</div>
+<div class="bar " role="banner">
+    <div class="bar__inner container">
+        <div class="ons-application-title bar__title venus"> SDX Console </div>&nbsp;
+    </div>
+</div>
+<div class="page__subheader"></div>
