+++ console/templates/partials/question.html
@@ -1,0 +1,16 @@
+{% macro build_question(title) %}
+<div class="question">
+    <h2>{{ title }}</h2>
+    <div class="question__answers">
+        <div class="question__answer">
+            <div class="answer answer--currency">
+                <div class="answer__fields js-fields">
+                    <div class="field">
+                        <input data-qa="input-text" class="input input--text input-type__input" id="input-text" value="" type="text" />
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endmacro %}
