+++ console/templates/partials/section.html
@@ -1,0 +1,7 @@
+{% macro section_default(title, description) -%}
+    <div class="section">
+        <h2 class="section__title saturn">{{ title }}</h2>
+        <p class="section__description mars">{{ description }}</p>
+        {{ caller() }}
+    </div>
+{% endmacro %}
