+++ console/templates/security/login_user.html
@@ -1,0 +1,22 @@
+{% import 'partials/section.html' as section %}
+{% extends "layouts/_twocol.html" %}
+
+{% block page_title %} Login {% endblock %}
+
+{% block content %}
+{% from "security/_macros.html" import render_field_with_errors, render_field %}
+{% include "security/_messages.html" %}
+
+  <h1>{{ _('Login') }}</h1>
+    <div class="top-pad">
+        <form action="{{ url_for_security('login') }}" method="POST" name="login_user_form">
+
+            <div class="form-group"> {{ render_field_with_errors(login_user_form.email) }} </div>
+            <div class="form-group"> {{ render_field_with_errors(login_user_form.password) }} </div>
+            {{ render_field(login_user_form.next) }}
+            {{ render_field(login_user_form.submit) }}
+        </form>
+
+    </div>
+  {% include "security/_menu.html" %}
+{% endblock content %}
