+++ console/templates/test.html
@@ -1,0 +1,11 @@
+{% import 'partials/section.html' as section %}
+{% extends "layouts/_twocol.html" %}
+
+{% block page_title %} Test Page {% endblock %}
+
+
+{% block content %}
+
+{% include 'partials/navigation_tabs.html' %}
+
+{% endblock content %}
